views

+++ quarantine/views.py
@@ -34,7 +34,7 @@ def login_view(request):
         if user is not None:
             # existe na BD
             login(request, user)
-            return menu(request)
+            return redirect('menu')
         #    return render(request, 'quarantine/menu.html')
         else:
             # nao existe na BD
@@ -46,7 +46,7 @@ def login_view(request):
 def logout_view(request):
     logout(request)
     # Redirect to a success page.
-    return menu(request)
+    return redirect('menu')
 
 
 # ------------------------------------------------------------------------
@@ -69,7 +69,7 @@ def registo_view(request):
     else:
         form = RegistrationForm()
         context['registration_form'] = form
-    return render(request, 'quarantine/registo.html', context)
+        return render(request, 'quarantine/registo.html', context)
 
 
 # ----------------------------------------------------------------------
@@ -103,23 +103,26 @@ def atualizarperfil(request, username):
 # ----------------------------------------------------------------------
 
 
-def criargrupopage(request):
-    users = User.objects.exclude(username=request.user.username)
-    return render(request, 'quarantine/criargrupopage.html', {'users': users})
+#def criargrupopage(request):
+#    users = User.objects.exclude(username=request.user.username)
+#    return render(request, 'quarantine/criargrupo.html', {'users': users})
 
 
 def criargrupo(request):
-    g = Grupo(titulo=request.POST['titulo'], descrição=request.POST['desc'])
-    g.save()
+    if request.POST:
+        g = Grupo(titulo=request.POST['titulo'], descrição=request.POST['desc'])
+        g.save()
 
-    mg = MembroGrupo(user=request.user, grupo=g, is_admin=True)
-    mg.save()
-    for username in request.POST.getlist('user'):
-        mg = MembroGrupo(user=User.objects.get(username=username), grupo=g, is_admin=False)
+        mg = MembroGrupo(user=request.user, grupo=g, is_admin=True)
         mg.save()
-    g.save()
-    return HttpResponseRedirect(reverse('menu', args=()))
-
+        for username in request.POST.getlist('user'):
+            mg = MembroGrupo(user=User.objects.get(username=username), grupo=g, is_admin=False)
+            mg.save()
+        g.save()
+        return HttpResponseRedirect(reverse('menu', args=()))
+    else:
+        users = User.objects.exclude(username=request.user.username)
+        return render(request, 'quarantine/criargrupo.html', {'users': users})
 
 def apagargrupo(request, grupo_id):
     grupo = get_object_or_404(Grupo, pk=grupo_id)
@@ -188,51 +191,64 @@ def apagarpublicacao(request, grupo_id, pub_id):
             "Não é admin do grupo ou autor da publicação!!"})
 
 
-def adicionarmembrospage(request, grupo_id):
-    grupo = get_object_or_404(Grupo, pk=grupo_id)
-    membros = MembroGrupo.objects.filter(grupo_id=grupo_id)
-    users = User.objects.exclude(id=request.user.id)
-    for membro in membros:
-        users = users.exclude(id=membro.user_id)
+#def adicionarmembrospage(request, grupo_id):
+    #grupo = get_object_or_404(Grupo, pk=grupo_id)
+    #membros = MembroGrupo.objects.filter(grupo_id=grupo_id)
+    #users = User.objects.exclude(id=request.user.id)
+    #   for membro in membros:
+
+
+#   users = users.exclude(id=membro.user_id)
 
     #    url = request.GET.get("next")
-    return render(request, 'quarantine/adicionarmembrospage.html', {'grupo': grupo, 'users': users,
-                                                                    # 'next': url
-                                                                    })
+    #    return render(request, 'quarantine/adicionarmembros.html', {'grupo': grupo, 'users': users,})
 
 
 def adicionarmembros(request, grupo_id):
     grupo = get_object_or_404(Grupo, pk=grupo_id)
+    if request.POST:
+
+        #    url = request.GET.get("next")
+        #    try:
+        #        resolve(url)
+        for username in request.POST.getlist('user'):
+            mg = MembroGrupo(user=User.objects.get(username=username), grupo=grupo, is_admin=False)
+            mg.save()
+        return HttpResponseRedirect(reverse('grupo_view', kwargs={"grupo_id": grupo_id}))
+    else:
+        membros = MembroGrupo.objects.filter(grupo_id=grupo_id)
+        users = User.objects.exclude(id=request.user.id)
+        for membro in membros:
+            users = users.exclude(id=membro.user_id)
+
+        #    url = request.GET.get("next")
+        return render(request, 'quarantine/adicionarmembros.html', {'grupo': grupo, 'users': users, })
+
+#def removermembrospage(request, grupo_id):
+    #grupo = get_object_or_404(Grupo, pk=grupo_id)
+    #users = grupo.membros.all()
 
     #    url = request.GET.get("next")
-    #    try:
-    #        resolve(url)
-    for username in request.POST.getlist('user'):
-        mg = MembroGrupo(user=User.objects.get(username=username), grupo=grupo, is_admin=False)
-        mg.save()
-    # return HttpResponseRedirect(reverse('grupo_view', args=grupo_id))
-    return grupo_view(request, grupo_id)
-
-
-def removermembrospage(request, grupo_id):
-    grupo = get_object_or_404(Grupo, pk=grupo_id)
-    users = grupo.membros.all()
-
-    #    url = request.GET.get("next")
-    return render(request, 'quarantine/removermembrospage.html', {'grupo': grupo, 'users': users,
+    #return render(request, 'quarantine/removermembros.html', {'grupo': grupo, 'users': users,
                                                                   # 'next': url
-                                                                  })
+#                                                              })
 
 
 def removermembros(request, grupo_id):
     grupo = get_object_or_404(Grupo, pk=grupo_id)
+    if request.POST:
 
-    for username in request.POST.getlist('user'):
-        mg = MembroGrupo.objects.get(user=User.objects.get(username=username), grupo=grupo)
-        mg.delete()
-    return grupo_view(request, grupo_id)
+        for username in request.POST.getlist('user'):
+            mg = MembroGrupo.objects.get(user=User.objects.get(username=username), grupo=grupo)
+            mg.delete()
+        return grupo_view(request, grupo_id)
+    else:
+        users = grupo.membros.all()
 
-
+        #    url = request.GET.get("next")
+        return render(request, 'quarantine/removermembros.html', {'grupo': grupo, 'users': users,
+                                                                      # 'next': url
+                                                                      })
 # ----------------------------------------------------------------------
